feat: store dedup state in issue tracker
This change stores the issue fingerprint in the issue being created
on Github. Prior to creating the issue a Github search is conducted
to ensure no issue already exists with this fingerprint. This avoids
having to use leveldb to deduplicate issues.

For my particular usecase that leverages Github actions this is more
desirable as it avoids needing to either commit the leveldb database
to the parent repository or pull it from an external source.